+++ example/lib/features/energy/presentation/energy_page.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+import '../../../shared/bloc/sankey_bloc.dart';
+import '../../../shared/widgets/sankey_page_scaffold.dart';
+import '../data/energy_repository.dart';
+
+class EnergyPage extends StatelessWidget {
+  const EnergyPage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (_) => SankeyBloc(const EnergyRepository()),
+      child: const SankeyPageScaffold(),
+    );
+  }
+}